обновление

+++ src/by/epamtc/lab1/task8/util/MathTool.java
@@ -1,13 +1,13 @@
 package by.epamtc.lab1.task8.util;
 
 public class MathTool {
-    public static void printIndexes(int masA[], int masB[]) {
+    public static void printIndexes(double masA[], double masB[]) {
         for (int i = 0; i < masB.length; i++) {
             System.out.print(binarySearch(masA, 0, masA.length - 1, masB[i])+" ");
         }
     }
 
-    static int binarySearch(int[] arr, int l, int r, int sElem) {
+    static double binarySearch(double[] arr, int l, int r, double sElem) {
         int m = -1;
         if (sElem <= arr[l])
             return l;
